base58: move into its own package
It doesn't belong to crypto in any way other than it uses hash function
internally.
