+++ pkg/encoding/base58/doc.go
@@ -1,0 +1,4 @@
+/*
+Package base58 wraps generic base58 encoder with NEO-specific checksumming.
+*/
+package base58
